Put Hazel back to a shared library.
Note:  There is no need to manually do #define tricks to get declspec(dllexport) to magically appear on windows.
cmake does this for you.

1) Make sure cmake is running with -DCMAKE_WINDOWS_EXPORT_ALL_SYMBOLS=TRUE"

2) Make sure the Hazel shared library doesnt contain any globals (like static class members) that need to be exported
(Apparently it is possible to get these to work via cmake's generate_export_header(), but I found it easier to just avoid having to export globals in the first place